+++ simple_live_app/lib/modules/category/detail/category_detail_controller.dart
