+++ simple_live_app/lib/modules/follow_user/follow_user_page.dart
@@ -1,0 +1,175 @@
+import 'package:flutter/material.dart';
+import 'package:get/get.dart';
+import 'package:remixicon/remixicon.dart';
+import 'package:simple_live_app/app/app_style.dart';
+import 'package:simple_live_app/app/sites.dart';
+import 'package:simple_live_app/modules/follow_user/follow_user_controller.dart';
+import 'package:simple_live_app/routes/app_navigation.dart';
+import 'package:simple_live_app/widgets/filter_button.dart';
+import 'package:simple_live_app/widgets/follow_user_item.dart';
+import 'package:simple_live_app/widgets/page_grid_view.dart';
+
+class FollowUserPage extends GetView<FollowUserController> {
+  const FollowUserPage({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    controller.filterMode.value = 0;
+    var count = MediaQuery.of(context).size.width ~/ 500;
+    if (count < 1) count = 1;
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text("关注用户"),
+        actions: [
+          PopupMenuButton(
+            itemBuilder: (context) {
+              return const [
+                PopupMenuItem(
+                  value: 0,
+                  child: Row(
+                    mainAxisSize: MainAxisSize.min,
+                    children: [
+                      Icon(Remix.save_2_line),
+                      AppStyle.hGap12,
+                      Text("导出文件")
+                    ],
+                  ),
+                ),
+                PopupMenuItem(
+                  value: 1,
+                  child: Row(
+                    mainAxisSize: MainAxisSize.min,
+                    children: [
+                      Icon(Remix.folder_open_line),
+                      AppStyle.hGap12,
+                      Text("导入文件")
+                    ],
+                  ),
+                ),
+                PopupMenuItem(
+                  value: 2,
+                  child: Row(
+                    mainAxisSize: MainAxisSize.min,
+                    children: [
+                      Icon(Remix.text),
+                      AppStyle.hGap12,
+                      Text("导出文本"),
+                    ],
+                  ),
+                ),
+                PopupMenuItem(
+                  value: 3,
+                  child: Row(
+                    mainAxisSize: MainAxisSize.min,
+                    children: [
+                      Icon(Remix.file_text_line),
+                      AppStyle.hGap12,
+                      Text("导入文本"),
+                    ],
+                  ),
+                ),
+              ];
+            },
+            onSelected: (value) {
+              if (value == 0) {
+                controller.exportFile();
+              } else if (value == 1) {
+                controller.inputFile();
+              } else if (value == 2) {
+                controller.exportText();
+              } else if (value == 3) {
+                controller.inputText();
+              }
+            },
+          ),
+        ],
+      ),
+      body: Column(
+        crossAxisAlignment: CrossAxisAlignment.stretch,
+        children: [
+          Padding(
+            padding: AppStyle.edgeInsetsL8,
+            child: Row(
+              children: [
+                Expanded(
+                  child: Obx(
+                    () => Wrap(
+                      spacing: 12,
+                      children: [
+                        FilterButton(
+                          text: "全部",
+                          selected: controller.filterMode.value == 0,
+                          onTap: () {
+                            controller.setFilterMode(0);
+                          },
+                        ),
+                        FilterButton(
+                          text: "直播中",
+                          selected: controller.filterMode.value == 1,
+                          onTap: () {
+                            controller.setFilterMode(1);
+                          },
+                        ),
+                        FilterButton(
+                          text: "未开播",
+                          selected: controller.filterMode.value == 2,
+                          onTap: () {
+                            controller.setFilterMode(2);
+                          },
+                        ),
+                      ],
+                    ),
+                  ),
+                ),
+                Obx(
+                  () => controller.updating.value
+                      ? TextButton.icon(
+                          onPressed: null,
+                          icon: const SizedBox(
+                            width: 16,
+                            height: 16,
+                            child: CircularProgressIndicator(
+                              strokeWidth: 2,
+                            ),
+                          ),
+                          label: const Text("更新状态中"),
+                        )
+                      : TextButton.icon(
+                          onPressed: () {
+                            controller.refreshData();
+                          },
+                          icon: const Icon(Icons.refresh),
+                          label: const Text("刷新"),
+                        ),
+                ),
+              ],
+            ),
+          ),
+          Expanded(
+            child: PageGridView(
+              crossAxisSpacing: 12,
+              crossAxisCount: count,
+              pageController: controller,
+              //firstRefresh: true,
+              showPCRefreshButton: false,
+              itemBuilder: (_, i) {
+                var item = controller.list[i];
+                var site = Sites.allSites[item.siteId]!;
+                return FollowUserItem(
+                  item: item,
+                  onRemove: () {
+                    controller.removeItem(item);
+                  },
+                  onTap: () {
+                    AppNavigator.toLiveRoomDetail(
+                        site: site, roomId: item.roomId);
+                  },
+                );
+              },
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+}
