+++ simple_live_app/lib/modules/settings/auto_exit_settings_page.dart
