+++ simple_live_app/lib/modules/settings/danmu_shield/danmu_shield_controller.dart
