+++ simple_live_app/lib/modules/settings/other/other_settings_controller.dart
