+++ simple_live_app/lib/modules/sync/local_sync/scan_qr/sync_scan_qr_controller.dart
