--- simple_live_app/lib/modules/toolbox/toolbox_controller.dart
@@ -1,179 +1,0 @@
-import 'package:dio/dio.dart';
-import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
-import 'package:flutter_smart_dialog/flutter_smart_dialog.dart';
-import 'package:get/get.dart';
-import 'package:simple_live_app/app/constant.dart';
-import 'package:simple_live_app/app/log.dart';
-import 'package:simple_live_app/app/sites.dart';
-import 'package:simple_live_app/routes/app_navigation.dart';
-
-class ToolBoxController extends GetxController {
-  final TextEditingController roomJumpToController = TextEditingController();
-  final TextEditingController getUrlController = TextEditingController();
-
-  void jumpToRoom(String e) async {
-    if (e.isEmpty) {
-      SmartDialog.showToast("链接不能为空");
-      return;
-    }
-    // 隐藏键盘
-    FocusManager.instance.primaryFocus?.unfocus();
-
-    var parseResult = await parse(e);
-    if (parseResult.isEmpty && parseResult.first == "") {
-      SmartDialog.showToast("无法解析此链接");
-      return;
-    }
-
-    // 延迟200ms跳转，等待键盘隐藏
-    Future.delayed(const Duration(milliseconds: 200), () {
-      Site site = parseResult[1];
-      AppNavigator.toLiveRoomDetail(site: site, roomId: parseResult.first);
-    });
-  }
-
-  void getPlayUrl(String e) async {
-    if (e.isEmpty) {
-      SmartDialog.showToast("链接不能为空");
-      return;
-    }
-    var parseResult = await parse(e);
-    if (parseResult.isEmpty && parseResult.first == "") {
-      SmartDialog.showToast("无法解析此链接");
-      return;
-    }
-    Site site = parseResult[1];
-    try {
-      SmartDialog.showLoading(msg: "");
-      var detail = await site.liveSite.getRoomDetail(roomId: parseResult.first);
-      var qualites = await site.liveSite.getPlayQualites(detail: detail);
-      SmartDialog.dismiss(status: SmartStatus.loading);
-      if (qualites.isEmpty) {
-        SmartDialog.showToast("读取直链失败,无法读取清晰度");
-
-        return;
-      }
-      var result = await Get.dialog(SimpleDialog(
-        title: const Text("选择清晰度"),
-        children: qualites
-            .map(
-              (e) => ListTile(
-                title: Text(
-                  e.quality,
-                  textAlign: TextAlign.center,
-                ),
-                onTap: () {
-                  Get.back(result: e);
-                },
-              ),
-            )
-            .toList(),
-      ));
-      if (result == null) {
-        return;
-      }
-      SmartDialog.showLoading(msg: "");
-      var playUrls =
-          await site.liveSite.getPlayUrls(detail: detail, quality: result);
-      SmartDialog.dismiss(status: SmartStatus.loading);
-      await Get.dialog(SimpleDialog(
-        title: const Text("选择线路"),
-        children: playUrls
-            .map(
-              (e) => ListTile(
-                title: Text(
-                  "线路${playUrls.indexOf(e) + 1}",
-                ),
-                subtitle: Text(
-                  e,
-                  maxLines: 1,
-                  overflow: TextOverflow.ellipsis,
-                ),
-                onTap: () {
-                  Clipboard.setData(ClipboardData(text: e));
-                  Get.back();
-                  SmartDialog.showToast("已复制直链");
-                },
-              ),
-            )
-            .toList(),
-      ));
-    } catch (e) {
-      SmartDialog.showToast("读取直链失败");
-    } finally {
-      SmartDialog.dismiss(status: SmartStatus.loading);
-    }
-  }
-
-  Future<List> parse(String url) async {
-    var id = "";
-    if (url.contains("bilibili.com")) {
-      var regExp = RegExp(r"bilibili\.com/([\d|\w]+)");
-      id = regExp.firstMatch(url)?.group(1) ?? "";
-      return [id, Sites.allSites[Constant.kBiliBili]!];
-    }
-
-    if (url.contains("b23.tv")) {
-      var btvReg = RegExp(r"https?:\/\/b23.tv\/[0-9a-z-A-Z]+");
-      var u = btvReg.firstMatch(url)?.group(0) ?? "";
-      var location = await getLocation(u);
-
-      return await parse(location);
-    }
-
-    if (url.contains("douyu.com")) {
-      var regExp = RegExp(r"douyu\.com/([\d|\w]+)");
-      id = regExp.firstMatch(url)?.group(1) ?? "";
-
-      return [id, Sites.allSites[Constant.kDouyu]!];
-    }
-    if (url.contains("huya.com")) {
-      var regExp = RegExp(r"huya\.com/([\d|\w]+)");
-      id = regExp.firstMatch(url)?.group(1) ?? "";
-
-      return [id, Sites.allSites[Constant.kHuya]!];
-    }
-    if (url.contains("live.douyin.com")) {
-      var regExp = RegExp(r"live\.douyin\.com/([\d|\w]+)");
-      id = regExp.firstMatch(url)?.group(1) ?? "";
-
-      return [id, Sites.allSites[Constant.kDouyin]!];
-    }
-    if (url.contains("webcast.amemv.com")) {
-      var regExp = RegExp(r"reflow/(\d+)");
-      id = regExp.firstMatch(url)?.group(1) ?? "";
-      return [id, Sites.allSites[Constant.kDouyin]!];
-    }
-    if (url.contains("v.douyin.com")) {
-      var regExp = RegExp(r"http.?://v.douyin.com/[\d\w]+/");
-      var u = regExp.firstMatch(url)?.group(0) ?? "";
-      var location = await getLocation(u);
-      return await parse(location);
-    }
-
-    return [];
-  }
-
-  Future<String> getLocation(String url) async {
-    try {
-      if (url.isEmpty) return "";
-      await Dio().get(
-        url,
-        options: Options(
-          followRedirects: false,
-        ),
-      );
-    } on DioException catch (e) {
-      if (e.response!.statusCode == 302) {
-        var redirectUrl = e.response!.headers.value("Location");
-        if (redirectUrl != null) {
-          return redirectUrl;
-        }
-      }
-    } catch (e) {
-      Log.logPrint(e);
-    }
-    return "";
-  }
-}
