--- simple_live_app/lib/modules/user/account/bilibili/web_login_page.dart
@@ -1,46 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:flutter_inappwebview/flutter_inappwebview.dart';
-import 'package:get/get.dart';
-import 'package:simple_live_app/modules/user/account/bilibili/web_login_controller.dart';
-
-class BiliBiliWebLoginPage extends GetView<BiliBiliWebLoginController> {
-  const BiliBiliWebLoginPage({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: const Text("哔哩哔哩账号登录"),
-        actions: [
-          TextButton.icon(
-            onPressed: controller.toQRLogin,
-            icon: const Icon(Icons.qr_code),
-            label: const Text("二维码登录"),
-          ),
-        ],
-      ),
-      body: InAppWebView(
-        onWebViewCreated: controller.onWebViewCreated,
-        onLoadStop: controller.onLoadStop,
-        initialOptions: InAppWebViewGroupOptions(
-          crossPlatform: InAppWebViewOptions(
-            userAgent:
-                "Mozilla/5.0 (iPhone; CPU iPhone OS 13_2_3 like Mac OS X) AppleWebKit/605.1.15 (KHTML, like Gecko) Version/13.0.3 Mobile/15E148 Safari/604.1 Edg/118.0.0.0",
-            useShouldOverrideUrlLoading: true,
-          ),
-        ),
-        shouldOverrideUrlLoading: (webController, navigationAction) async {
-          var uri = navigationAction.request.url;
-          if (uri == null) {
-            return NavigationActionPolicy.ALLOW;
-          }
-          if (uri.host == "m.bilibili.com" || uri.host == "www.bilibili.com") {
-            await controller.logined();
-            return NavigationActionPolicy.CANCEL;
-          }
-          return NavigationActionPolicy.ALLOW;
-        },
-      ),
-    );
-  }
-}
